refactor(shared): 归还提示词所有权并收紧边界
